Condensed the way img have been added to <li>
Rather than having the image on a separate line to create the difference between the text and the image, I have put the image code within the list item element and used <br> to create the spaces. I think this is a little more condensed than before. I'm not sure if it more efficient though.